policy + gate

+++ app/Http/Controllers/ArticleController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Models\Article;
 use App\Models\Comment;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class ArticleController extends Controller
 {
@@ -22,6 +23,7 @@ class ArticleController extends Controller
      */
     public function create()
     {
+        Gate::authorize('create', Article::class);
         return view('article/create');
     }
 
@@ -30,6 +32,7 @@ class ArticleController extends Controller
      */
     public function store(Request $request)
     {
+        Gate::authorize('create', Article::class);
         $request->validate([
             'title'=>'required|min:10',
             'text'=>'required|max:300'
@@ -57,6 +60,7 @@ class ArticleController extends Controller
      */
     public function edit(Article $article)
     {
+        Gate::authorize('update', $article);
         return view('article/edit', ['article'=>$article]);
     }
 
@@ -65,6 +69,7 @@ class ArticleController extends Controller
      */
     public function update(Request $request, Article $article)
     {
+        Gate::authorize('update', $article);
         $request->validate([
             'title'=>'required|min:10',
             'text'=>'required|max:300'
@@ -82,6 +87,7 @@ class ArticleController extends Controller
      */
     public function destroy(Article $article)
     {
+        Gate::authorize('delete', $article, Article::class);
         Comment::where('articles_id', $article->id)->delete();
         $article->delete();
         return redirect()->route('article.index')->with('success', 'Статья и все комментарии удалены');
